requirements.txt
altair==4.2.0
pandas==1.5.0
Pillow==9.2.0
streamlit==1.15.2
